support for custom render bounding boxes + misc fixes

+++ src/main/java/com/simibubi/create/foundation/fluid/FluidIngredient.java
@@ -56,7 +56,7 @@ public abstract class FluidIngredient implements Predicate<FluidStack> {
 		return ingredient;
 	}
 
-	protected int amountRequired;
+	protected long amountRequired;
 
 	protected abstract boolean testInternal(FluidStack t);
